feat: Cross-platform network IP detection and check script update
Enhanced Network.java to dynamically determine the default LAN IP address using a robust method, replacing the static '0.0.0.0' value. Updated check.sh to use Checkstyle v12.0.1 and added a Windows-compatible check.bat script. Improved documentation in CONTRIBUTING.md and fixed minor formatting in CODE_OF_CONDUCT.md. Updated build.gradle with platform version info and replaced checkstyle.jar binary.

This is a step forward in allowing the software to be developed in multiple-platforms (specifically targeting Windows this time). This fixes a network bug with the default IP configuration (0.0.0.0) which is undefined on Windows, instead we now grab the user's IP address and set that as the default which works for all platforms.