feat: Implement Edit Interpreter
feat: Implement Edit Interpreter

Created EditInterpreter for consolidating and applying edit packets:
  - Track pending edits per file with automatic merging of consecutive edits
  - Apply finalized edits to file line lists
  - Process edits in reverse order for correct line number preservation
EditSerializer was set to final for consistency and since it will not be extended by any other classes. FileComparer has the ability to compare lines from two lists of strings (to be used for the FileHandler to avoid reading the same lines from a file).

This will allow us to interpret the edit packets received by the server and actually be able to apply and write them to our files!