feat: Implement networked file creation
feat: Implement networked file creation

FileHandler allows for file creation via the Files.createFile() method and sends file creation packets similar to how the onModify() method sends packets. ClientHandler stores a set of Strings marking the relative files that will be created. The Network registers the CreateFilePacket for both the client and the server. The Packet class features a new CreateFilePacket class structured similarly to the DeleteFilePacket (simply storing the file name). The ProjectScene checks for a CreateFilePacket and creates the specified file using the FileHandler.

We can now create, send changes, and delete files through the network! Directory support is the next step as the current system is set up to only work for text files.